requirements.txt
numpy
pandas
requests
tiktoken
openai
azure-ai-projects
azure-ai-agents
azure-identity
chromadb
tqdm
torch
dftest
scipy
